+++ src/lib/getActivationStart.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 Google LLC
+ * Copyright 2022 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,22 +14,10 @@
  * limitations under the License.
  */
 
+import {getNavigationEntry} from './getNavigationEntry.js';
 
-/**
- * Returns a promise that resolves once the browser window has loaded and
- * all load callbacks have finished executing.
- * @return {Promise<void>}
- */
-function afterLoad() {
-  return browser.executeAsync((done) => {
-    if (document.readyState === 'complete') {
-      setTimeout(done, 0);
-    } else {
-      addEventListener('load', () => setTimeout(done, 0));
-    }
-  });
-}
 
-module.exports = {
-  afterLoad,
+export const getActivationStart = (): number => {
+  const navEntry = getNavigationEntry();
+  return navEntry && navEntry.activationStart || 0;
 };